Add dashboard (without table) (#39)
* Add basic dahboard (not done yet)

* Add working top of the dashboard

* Add fully funtional dashboard without table

* Fix the database fields

* Add text if no data

* Fix Income category in outcome analysis

* Fix dashboard logic

* Fix progress bar over 100%

+++ app/database/transactions.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Double, Date, Sequence, ForeignKey
+from sqlalchemy import Column, Integer, String, Double, Date, Sequence, ForeignKey, Numeric
 
 
 def create_transactions_table(db):
@@ -6,7 +6,7 @@ def create_transactions_table(db):
         __tablename__ = "transactions"
         id = Column(Integer, Sequence("seq_transactions_id", start=1), primary_key=True)
         transaction_date = Column(Date, nullable=False)
-        value = Column(Double, nullable=False)
+        value = Column(Numeric(10, 2), nullable=False)
         category_id = Column(Integer, ForeignKey("category.id"), nullable=True)
         subcategory_id = Column(Integer, ForeignKey("subcategory.id"), nullable=True)
         goal_id = Column(Integer, ForeignKey("goals.id"), nullable=True)
